Publish messages to specific exchange
We needed a rabbitmq.config file to configure the exchange that mqtt
clients will publish too. Otherwise, they publish to amq.topic by default.
This file is currently the only way to configure plugins in rabbit.

Adding env vars to docker-compose was overwriting the rabbitmq.config
file that we were copying into the container. This is why we define the
username and password in the Dockerfile now.